+++ common/src/main/java/dev/ftb/mods/ftbquests/quest/theme/property/StringProperty.java
@@ -1,8 +1,5 @@
 package dev.ftb.mods.ftbquests.quest.theme.property;
 
-/**
- * @author LatvianModder
- */
 public class StringProperty extends ThemeProperty<String> {
 	public StringProperty(String n) {
 		super(n, "");