+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2101.1.18]
+
+### Added
+* Added configurable repeat cooldowns for repeatable quests
+  * See the new "Quest Repeat Cooldown" field in the Misc section of the quests properties screen
+  * Repeat cooldown is triggered when the first player on the team claims the quest reward for the quest
+* More visually attractive highlighting of the currently-selected chapter
+  * Also added two new fields to `ftb_quest_theme.txt` resource pack file: `selected_chapter_highlight_1` and `selected_chapter_highlight_2`
+
 ## [2101.1.17]
 
 ### Added